refactor : 프로젝트 구조변경 및 Strategy패턴 적용
