+++ src/main/java/racing/Application.java
@@ -1,5 +1,7 @@
 package racing;
 
+import racing.controller.Game;
+
 public class Application {
     public static void main(String[] args) {
         Game game = new Game();
