+++ src/main/java/racing/strategy/GoStraightStrategy.java
@@ -1,0 +1,7 @@
+package racing.strategy;
+
+@FunctionalInterface
+public interface GoStraightStrategy {
+
+    boolean isGoStraight();
+}
